+++ mayan/apps/documents/tests/test_document_page_views.py
@@ -132,6 +132,30 @@ class DocumentPageViewTestCase(
             response=response, status_code=200, text=self.test_document.label
         )
 
+    def test_document_page_rotate_left_view_no_permission(self):
+        response = self._request_test_document_page_rotate_left_view()
+        self.assertEqual(response.status_code, 404)
+
+    def test_document_page_rotate_left_view_with_access(self):
+        self.grant_access(
+            obj=self.test_document, permission=permission_document_view
+        )
+
+        response = self._request_test_document_page_rotate_left_view()
+        self.assertEqual(response.status_code, 302)
+
+    def test_document_page_rotate_right_view_no_permission(self):
+        response = self._request_test_document_page_rotate_right_view()
+        self.assertEqual(response.status_code, 404)
+
+    def test_document_page_rotate_right_view_with_access(self):
+        self.grant_access(
+            obj=self.test_document, permission=permission_document_view
+        )
+
+        response = self._request_test_document_page_rotate_right_view()
+        self.assertEqual(response.status_code, 302)
+
     def test_document_page_view_no_permissions(self):
         response = self._request_test_document_page_view(
             document_page=self.test_document.pages.first()
@@ -151,3 +175,27 @@ class DocumentPageViewTestCase(
                 self.test_document.pages.first()
             )
         )
+
+    def test_document_page_zoom_in_view_no_permission(self):
+        response = self._request_test_document_page_zoom_in_view()
+        self.assertEqual(response.status_code, 404)
+
+    def test_document_page_zoom_in_view_with_access(self):
+        self.grant_access(
+            obj=self.test_document, permission=permission_document_view
+        )
+
+        response = self._request_test_document_page_zoom_in_view()
+        self.assertEqual(response.status_code, 302)
+
+    def test_document_page_zoom_out_view_no_permission(self):
+        response = self._request_test_document_page_zoom_out_view()
+        self.assertEqual(response.status_code, 404)
+
+    def test_document_page_zoom_out_view_with_access(self):
+        self.grant_access(
+            obj=self.test_document, permission=permission_document_view
+        )
+
+        response = self._request_test_document_page_zoom_out_view()
+        self.assertEqual(response.status_code, 302)